parallelize nstemplatetier test
